requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
scikit-learn==1.3.0
fastapi==0.103.1
uvicorn==0.23.2
evidently==0.4.5
joblib==1.3.2
sqlalchemy==2.0.21
pandas==2.2.2
numpy==1.26.4
requests==2.31.0